+++ tests/DependencyInjection/MercureExtensionTest.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Symfony\Bundle\MercureBundle\Tests\DependencyInjection;
 
+use Lcobucci\JWT\Signer\Key;
 use PHPUnit\Framework\TestCase;
 use Symfony\Bundle\MercureBundle\DependencyInjection\MercureExtension;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -20,6 +21,8 @@ use Symfony\Component\DependencyInjection\ParameterBag\ParameterBag;
 use Symfony\Component\Mercure\FrankenPhpHub;
 use Symfony\Component\Mercure\Hub;
 use Symfony\Component\Mercure\HubRegistry;
+use Symfony\Component\Mercure\Jwt\LcobucciFactory;
+use Symfony\Component\Mercure\ProtocolVersion;
 
 /**
  * @author Kévin Dunglas <dunglas@gmail.com>
@@ -48,6 +51,8 @@ class MercureExtensionTest extends TestCase
         $this->assertArrayHasKey('mercure.publisher', $container->getDefinition('mercure.hub.default.publisher')->getTags());
         $this->assertSame($config['mercure']['hubs']['default']['url'], $container->getDefinition('mercure.hub.default')->getArgument(0));
         $this->assertSame($config['mercure']['hubs']['default']['jwt'], $container->getDefinition('mercure.hub.default.jwt.provider')->getArgument(0));
+        $this->assertSame(ProtocolVersion::Legacy, $container->getDefinition('mercure.hub.default')->getArgument(5));
+        $this->assertNull($container->getDefinition('mercure.hub.default')->getArgument(6));
 
         $this->assertArrayHasKey('Symfony\Component\Mercure\HubInterface $default', $container->getAliases());
         $this->assertArrayHasKey('Symfony\Component\Mercure\PublisherInterface $default', $container->getAliases());
@@ -103,9 +108,12 @@ class MercureExtensionTest extends TestCase
         $this->assertTrue($container->hasDefinition('mercure.hub.managed.publisher')); // Publisher
         $this->assertTrue($container->hasDefinition('mercure.hub.managed.jwt.provider'));
         $this->assertTrue($container->hasDefinition('mercure.hub.managed.jwt.factory'));
+        $this->assertSame(LcobucciFactory::class, $container->getDefinition('mercure.hub.managed.jwt.factory')->getClass());
         $this->assertArrayHasKey('mercure.publisher', $container->getDefinition('mercure.hub.managed.publisher')->getTags());
         $this->assertSame($config['mercure']['hubs']['managed']['url'], $container->getDefinition('mercure.hub.managed')->getArgument(0));
         $this->assertSame($config['mercure']['hubs']['managed']['jwt']['secret'], $container->getDefinition('mercure.hub.managed.jwt.factory')->getArgument(0));
+        $this->assertSame(ProtocolVersion::Legacy, $container->getDefinition('mercure.hub.managed')->getArgument(5));
+        $this->assertNull($container->getDefinition('mercure.hub.managed')->getArgument(6));
         $this->assertSame([$config['mercure']['hubs']['managed']['jwt']['subscribe']], $container->getDefinition('mercure.hub.managed.jwt.provider')->getArgument(1));
         $this->assertSame($config['mercure']['hubs']['managed']['jwt']['publish'], $container->getDefinition('mercure.hub.managed.jwt.provider')->getArgument(2));
 
@@ -133,6 +141,8 @@ class MercureExtensionTest extends TestCase
         $this->assertSame($config['mercure']['hubs']['managed2']['jwt']['passphrase'], $container->getDefinition('mercure.hub.managed2.jwt.factory')->getArgument(3));
         $this->assertSame([$config['mercure']['hubs']['managed2']['jwt']['subscribe']], $container->getDefinition('mercure.hub.managed2.jwt.provider')->getArgument(1));
         $this->assertSame($config['mercure']['hubs']['managed2']['jwt']['publish'], $container->getDefinition('mercure.hub.managed2.jwt.provider')->getArgument(2));
+        $this->assertSame(ProtocolVersion::Legacy, $container->getDefinition('mercure.hub.managed2')->getArgument(5));
+        $this->assertNull($container->getDefinition('mercure.hub.managed2')->getArgument(6));
 
         $this->assertArrayHasKey('Symfony\Component\Mercure\HubInterface $managed2', $container->getAliases());
         $this->assertArrayHasKey('Symfony\Component\Mercure\PublisherInterface $managed2', $container->getAliases());
@@ -155,6 +165,8 @@ class MercureExtensionTest extends TestCase
         $this->assertSame($config['mercure']['hubs']['demo']['url'], $container->getDefinition('mercure.hub.demo')->getArgument(0));
         $this->assertSame($config['mercure']['hubs']['demo']['public_url'], $container->getDefinition('mercure.hub.demo')->getArgument(3));
         $this->assertSame($config['mercure']['hubs']['demo']['jwt']['value'], $container->getDefinition('mercure.hub.demo.jwt.provider')->getArgument(0));
+        $this->assertSame(ProtocolVersion::Legacy, $container->getDefinition('mercure.hub.demo')->getArgument(5));
+        $this->assertNull($container->getDefinition('mercure.hub.demo')->getArgument(6));
 
         $this->assertArrayHasKey('Symfony\Component\Mercure\HubInterface $demo', $container->getAliases());
         $this->assertArrayHasKey('Symfony\Component\Mercure\PublisherInterface $demo', $container->getAliases());
@@ -211,6 +223,8 @@ class MercureExtensionTest extends TestCase
         $this->assertSame($config['mercure']['hubs'][0]['url'], $registry->getHub()->getUrl());
         $this->assertSame($config['mercure']['hubs'][0]['url'], $registry->getHub('default')->getUrl());
         $this->assertSame($config['mercure']['hubs'][1]['url'], $registry->getHub('managed')->getUrl());
+        $this->assertSame(ProtocolVersion::Legacy, $registry->getHub()->getProtocolVersion());
+        $this->assertSame('mercureAuthorization', $registry->getHub()->getCookieName());
     }
 
     public function testExtensionBuiltin()
@@ -235,6 +249,143 @@ class MercureExtensionTest extends TestCase
 
         $this->assertTrue($container->hasDefinition('mercure.hub.default'));
         $this->assertSame(FrankenPhpHub::class, $container->getDefinition('mercure.hub.default')->getClass());
+        $this->assertSame(ProtocolVersion::Legacy, $container->getDefinition('mercure.hub.default')->getArgument(2));
+        $this->assertNull($container->getDefinition('mercure.hub.default')->getArgument(3));
+    }
+
+    public function testExtensionProtocolVersion1_0()
+    {
+        $config = [
+            'mercure' => [
+                'hubs' => [
+                    'default' => [
+                        'url' => 'https://demo.mercure.rocks/hub',
+                        'jwt' => 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.e30.HB0k08BaV8KlLZ3EafCRlTDGbkd9qdznCzJQ_l8ELTU',
+                        'protocol_version' => '1.0',
+                        'cookie_name' => 'custom_cookie',
+                    ],
+                ],
+            ],
+        ];
+
+        $container = new ContainerBuilder(new ParameterBag(['kernel.debug' => false]));
+        (new MercureExtension())->load($config, $container);
+
+        $this->assertSame(ProtocolVersion::V1, $container->getDefinition('mercure.hub.default')->getArgument(5));
+        $this->assertSame('custom_cookie', $container->getDefinition('mercure.hub.default')->getArgument(6));
+    }
+
+    public function testLegacyHubUsesLcobucciFactoryWithLegacyProtocolVersion()
+    {
+        $config = [
+            'mercure' => [
+                'hubs' => [
+                    'default' => [
+                        'url' => 'https://demo.mercure.rocks/hub',
+                        'jwt' => [
+                            'secret' => '!ChangeMe!',
+                        ],
+                    ],
+                ],
+            ],
+        ];
+
+        $container = new ContainerBuilder(new ParameterBag(['kernel.debug' => false]));
+        (new MercureExtension())->load($config, $container);
+
+        $definition = $container->getDefinition('mercure.hub.default.jwt.factory');
+        $this->assertSame(LcobucciFactory::class, $definition->getClass());
+        $this->assertSame('!ChangeMe!', $definition->getArgument(0));
+        $this->assertSame(ProtocolVersion::Legacy, $definition->getArgument(4));
+    }
+
+    public function testV1HubUsesLcobucciFactoryWithV1ProtocolVersion()
+    {
+        $config = [
+            'mercure' => [
+                'hubs' => [
+                    'default' => [
+                        'url' => 'https://demo.mercure.rocks/hub',
+                        'jwt' => [
+                            'secret' => '!ChangeMe!',
+                        ],
+                        'protocol_version' => '1.0',
+                    ],
+                ],
+            ],
+        ];
+
+        $container = new ContainerBuilder(new ParameterBag(['kernel.debug' => false]));
+        (new MercureExtension())->load($config, $container);
+
+        $definition = $container->getDefinition('mercure.hub.default.jwt.factory');
+        $this->assertSame(LcobucciFactory::class, $definition->getClass());
+        $this->assertSame('!ChangeMe!', $definition->getArgument(0));
+        $this->assertSame(ProtocolVersion::V1, $definition->getArgument(4));
+    }
+
+    public function testV1HubWithSecretAndClaimsMintsATokenEndToEnd()
+    {
+        if (!class_exists(Key\InMemory::class)) {
+            $this->markTestSkipped('requires lcobucci/jwt.');
+        }
+
+        $config = [
+            'mercure' => [
+                'hubs' => [
+                    'default' => [
+                        'url' => 'https://demo.mercure.rocks/hub',
+                        'jwt' => [
+                            'secret' => '!ChangeMe!ChangeMe!ChangeMe!ChangeMe!',
+                            'claims' => ['iss' => 'https://example.com', 'sub' => 'https://example.com', 'client_id' => 'https://example.com'],
+                        ],
+                        'protocol_version' => '1.0',
+                    ],
+                ],
+            ],
+        ];
+
+        $container = new ContainerBuilder(new ParameterBag(['kernel.debug' => false]));
+        (new MercureExtension())->load($config, $container);
+
+        $container->getDefinition(HubRegistry::class)->setPublic(true);
+        $container->compile();
+
+        // The whole point of this test: FactoryTokenProvider must actually forward the "claims"
+        // (plus the auto-derived "aud") down to WebTokenFactory, or this throws.
+        $jwt = $container->get(HubRegistry::class)->getHub()->getProvider()->getJwt();
+
+        $this->assertMatchesRegularExpression('/^[\w-]+\.[\w-]+\.[\w-]+$/', $jwt);
+    }
+
+    public function testV1HubDefaultsAudienceToTheHubUrl()
+    {
+        if (!class_exists(Key\InMemory::class)) {
+            $this->markTestSkipped('requires lcobucci/jwt.');
+        }
+
+        $config = [
+            'mercure' => [
+                'hubs' => [
+                    'default' => [
+                        'url' => 'https://demo.mercure.rocks/hub',
+                        'jwt' => [
+                            'secret' => '!ChangeMe!ChangeMe!ChangeMe!ChangeMe!',
+                            'claims' => ['iss' => 'https://example.com', 'sub' => 'https://example.com', 'client_id' => 'https://example.com'],
+                        ],
+                        'protocol_version' => '1.0',
+                    ],
+                ],
+            ],
+        ];
+
+        $container = new ContainerBuilder(new ParameterBag(['kernel.debug' => false]));
+        (new MercureExtension())->load($config, $container);
+
+        $this->assertSame(
+            ['iss' => 'https://example.com', 'sub' => 'https://example.com', 'client_id' => 'https://example.com', 'aud' => 'https://demo.mercure.rocks/hub'],
+            $container->getDefinition('mercure.hub.default.jwt.provider')->getArgument(3)
+        );
     }
 }